split up components of code editors into separate classes
